Update docs and fix tests
