+++ .github/workflows/publish.yml
@@ -14,7 +14,7 @@ jobs:
             pull-requests: none
 
         steps:
-            - uses: actions/checkout@v4
+            - uses: actions/checkout@v5
               with:
                   fetch-depth: 0
                   persist-credentials: true
