+++ AGENTS.md
@@ -10,14 +10,14 @@ Welcome! This project is a Bun-first CLI that orchestrates transcription jobs th
 - `tsdown.config.ts` – Source of truth for bundling. `tsdown` reads this file directly, so keep entrypoints, externals and banners in sync with CLI expectations.
 - `README.md` – Landing page for the CLI. Update feature lists, flag descriptions and workflow notes whenever behaviour changes.
 - `AGENTS.md` – This document. Extend it when you add tooling, conventions or new directories.
-- `bun.lock` – Bun’s lockfile. Commit changes after running `bun update --latest` or touching dependencies.
+- `bun.lock` – Bun's lockfile. Commit changes after running `bun update --latest` or touching dependencies.
 - `biome.json` – Formatting/linting rules. `$schema` must match the currently installed Biome version.
 
 ## Tooling & workflow
 
-- **Package management:** Bun only. Run `bun install` after editing dependencies. We pin `https-proxy-agent@5.0.1` via overrides to keep `ffmpeg-static` installs happy behind proxies and force `env-paths@2.2.1` so the `ffmpeg-static` install script can run in CommonJS mode—leave those overrides alone unless you know what you’re doing.
+- **Package management:** Bun only. Run `bun install` after editing dependencies.
 - **Dependency updates:** Use `bun update --latest` to stay on the newest stack, then commit both `package.json` and `bun.lock`.
-- **Build:** `bun run build` shells out to `bunx tsdown --config tsdown.config.ts`. `tsdown` injects the Bun shebang via the `banner` config and the `postbuild` script sets the executable bit on `dist/index.js`. Preserve both behaviours whenever you touch build scripts.
+- **Build:** `bun run build` shells out to `bunx tsdown` using `tsdown.config.ts`. The bundler injects the Node.js shebang via the `banner` config. Preserve this behaviour when touching build scripts.
 - **Testing:** `bun test`. Add coverage alongside any new exported helper. Mock networked services (tafrigh, YouTube, Facebook, X/Twitter) so the suite runs offline.
 - **Lint/format:** Use Biome through `bun run lint` / `bun run format`. Keep `$schema` on the latest release and lint `src` plus `tsdown.config.ts` only (generated `dist/` stays excluded).
 
@@ -27,7 +27,15 @@ Welcome! This project is a Bun-first CLI that orchestrates transcription jobs th
 - Prefer explicit named exports and avoid defaults in `src/utils` for easier mocking.
 - Keep log messages actionable and leverage the existing `pino` logger utilities instead of `console` inside the CLI flow.
 - Avoid `try/catch` around imports per repo guidelines; use targeted error handling elsewhere.
-- Tests should describe behaviour (`it('should ...')`), assert against real return values, and isolate filesystem/network activity with Bun’s mocking tools.
+- Tests should describe behaviour (`it('should ...')`), assert against real return values, and isolate filesystem/network activity with Bun's mocking tools.
+
+## Media processing
+
+The CLI uses `youtube-dl-exec` (a wrapper around yt-dlp) for YouTube video/playlist downloads. This requires:
+- **Python 3.14+** available as `python3` in your system PATH
+- The package auto-installs yt-dlp during `bun install`, but you can also use a global `yt-dlp` binary by configuring `youtube-dl-exec` with `create('yt-dlp')`
+
+When testing `mediaUtils.ts`, ensure mocks are defined via `mock.module()` **before** importing the module to prevent real network calls.
 
 ## Documentation
