+++ bunfig.toml
@@ -1,0 +1,2 @@
+[test]
+preload = ["./test-setup.ts"]
